Completed 4.14-4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -1,21 +1,82 @@
-const notesRouter = require('express').Router()
+const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
-notesRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
-notesRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
+
+blogsRouter.post('/', async (request, response, next) => {
+
+  const body = request.body
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+  } catch (error) {
+    if (error.name === "TokenExpiredError") {
+      return response.status(401).json({ error: 'token expired' })
+    } else if (error.name === "JsonWebTokenError") {
+      return response.status(401).json({ error: 'invalid token'})
+    }
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+
+  })
   const savedBlog = await blog.save()
+    .catch(error => next(error))
+    return response.end()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog.toJSON())
 
 })
 
-notesRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  const userid = decodedToken.id
+  if ( blog.user.toString() === userid.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'authorization failed'})
+  }
+
 })
 
-module.exports = notesRouter
+blogsRouter.put('/:id', async (request, response) => {
+  const body = request.body
+
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, { likes: body.likes }, { new: true })
+
+  response.status(200).json(updatedBlog.toJSON())
+
+})
+
+module.exports = blogsRouter
